refactor: move API key validation component to be reusable
