+++ gravitee-apim-console-webui/src/components/api-key-validation/api-key-validation.component.html
