+++ gravitee-apim-console-webui/src/management/api/subscriptions/api-subscriptions.module.ts
@@ -48,13 +48,13 @@ import { ApiSubscriptionEditComponent } from './edit/api-subscription-edit.compo
 import { ApiSubscriptionListComponent } from './list/api-subscription-list.component';
 import { ApiPortalSubscriptionChangeEndDateDialogComponent } from './components/dialogs/change-end-date/api-portal-subscription-change-end-date-dialog.component';
 import { ApiPortalSubscriptionValidateDialogComponent } from './components/dialogs/validate/api-portal-subscription-validate-dialog.component';
-import { ApiKeyValidationComponent } from './components/api-key-validation/api-key-validation.component';
 import { ApiPortalSubscriptionRejectDialogComponent } from './components/dialogs/reject/api-portal-subscription-reject-dialog.component';
 import { ApiPortalSubscriptionRenewDialogComponent } from './components/dialogs/renew/api-portal-subscription-renew-dialog.component';
 import { ApiPortalSubscriptionExpireApiKeyDialogComponent } from './components/dialogs/expire-api-key/api-portal-subscription-expire-api-key-dialog.component';
 
 import { GioTableWrapperModule } from '../../../shared/components/gio-table-wrapper/gio-table-wrapper.module';
 import { GioPermissionModule } from '../../../shared/components/gio-permission/gio-permission.module';
+import { ApiKeyValidationModule } from '../../../components/api-key-validation/api-key-validation.module';
 
 @NgModule({
   declarations: [
@@ -66,13 +66,11 @@ import { GioPermissionModule } from '../../../shared/components/gio-permission/g
     ApiPortalSubscriptionRejectDialogComponent,
     ApiPortalSubscriptionRenewDialogComponent,
     ApiPortalSubscriptionValidateDialogComponent,
-
     ApiPortalSubscriptionExpireApiKeyDialogComponent,
-
-    ApiKeyValidationComponent,
   ],
-  exports: [ApiSubscriptionEditComponent, ApiSubscriptionListComponent, ApiKeyValidationComponent],
+  exports: [ApiSubscriptionEditComponent, ApiSubscriptionListComponent],
   imports: [
+    ApiKeyValidationModule,
     CommonModule,
     FormsModule,
     ReactiveFormsModule,
